UserService almost complete Added ReviewModel.cs

+++ Squids-Movies-App/SquidsMovieApp.DTO/UserModel.cs
@@ -19,7 +19,7 @@ namespace SquidsMovieApp.DTO
         public string Email { get; set; }
         public string Password { get; set; }
         public bool IsAdmin { get; set; }
-        public int MoneyBalance { get; set; }
+        public decimal MoneyBalance { get; set; }
         public ICollection<Participant> LikedActors { get; set; }
         public ICollection<Participant> LikedDirectors { get; set; }
         public ICollection<Movie> LikedMovies { get; set; }
